requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
asyncpg==0.29.0
python-dateutil==2.8.2
